fix: should_return_no_content_when_delete_given_id

+++ src/main/java/com/rest/springbootemployee/service/CompanyService.java
@@ -58,9 +58,12 @@ public class CompanyService {
     public void deleteCompanyById(int id) {
         Company company = findById(id);
         List<Employee> employees = company.getEmployees();
-        employees.forEach(employee -> {
-            jpaEmployeeRepository.deleteById(employee.getId());
-        });
+        if (employees != null){
+            employees.forEach(employee -> {
+                jpaEmployeeRepository.deleteById(employee.getId());
+            });
+        }
+
         companyRepository.deleteById(id);
     }
